Introduce Makefile and make mono happy
