+++ DATReaderTest/DATReaderTest.csproj
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <Product>Tester</Product>
     <Copyright>Copyright © 2017</Copyright>
   </PropertyGroup>
