+++ ROMVault/ROMVault.csproj
@@ -57,7 +57,7 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
     <DebugSymbols>true</DebugSymbols>
-    <OutputPath>E:\RomVaultTDC\</OutputPath>
+    <OutputPath>..\..\Stage</OutputPath>
     <DefineConstants>TRACE;DEBUG</DefineConstants>
     <DebugType>full</DebugType>
     <PlatformTarget>x64</PlatformTarget>
@@ -152,7 +152,7 @@
     <Compile Include="FrmRegistration.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="FrmRegistration.designer.cs">
+    <Compile Include="FrmRegistration.Designer.cs">
       <DependentUpon>FrmRegistration.cs</DependentUpon>
     </Compile>
     <Compile Include="Report.cs" />
