+++ RVCmd/RVCmd.csproj
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <Version>3.1.2</Version>
     <FileVersion>1.0.0.0</FileVersion>
     <Copyright>Copyright © 2020</Copyright>
