+++ RVWorld.sln
@@ -7,7 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrrntZipUI", "TrrntZipUI\Tr
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RomVaultX", "RomVaultX\RomVaultX.csproj", "{B5A4FB1B-FF88-4EB8-95F0-BCE4ED52E020}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RomVault", "ROMVault\RomVault.csproj", "{437E12AB-46F3-4A75-A3D5-6C524B5A79B5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RomVault", "ROMVault\ROMVault.csproj", "{437E12AB-46F3-4A75-A3D5-6C524B5A79B5}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Trrntzip", "Trrntzip\Trrntzip.csproj", "{A08C2584-77FF-459A-AB2B-2E3042818A8E}"
 EndProject
