+++ TrrntZipCMD/TrrntZipCMD.csproj
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <AssemblyVersion>1.0.5.0</AssemblyVersion>
     <Copyright>Copyright © 2020</Copyright>
   </PropertyGroup>
