+++ TrrntZipUI/TrrntZipUI.csproj
@@ -68,7 +68,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
-    <Reference Include="System.configuration" />
+    <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.Windows.Forms" />
@@ -80,7 +80,7 @@
     <Compile Include="FrmTrrntzip.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="FrmTrrntzip.designer.cs">
+    <Compile Include="FrmTrrntzip.Designer.cs">
       <DependentUpon>FrmTrrntzip.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
